+++ src/main/java/com/BestBot/Core/Configuration/Listeners/MessageCreateListener.java
@@ -1,17 +1,13 @@
-package com.BestBot.Core.Listeners;
+package com.BestBot.Core.Configuration.Listeners;
 
-import org.springframework.stereotype.Service;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import net.dv8tion.jda.api.requests.restaction.MessageAction;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Mono;
 
-@Service
-@Configuration
 @Component
 public class MessageCreateListener extends ListenerAdapter {
