--- src/main/java/com/BestBot/Core/Parsers/CrossoutdbParser.java
@@ -1,12 +1,0 @@
-package com.BestBot.Core.Parsers;
-
-import com.BestBot.Core.Repository.ItemEntityRepository;
-import com.fasterxml.jackson.databind.JsonNode;
-
-import java.util.List;
-
-public interface CrossoutdbParser {
-    String getApi();
-//    List<Object> testTest();
-
-}
