+++ src/main/java/com/BestBot/Core/Service/CrossoutdbParser.java
@@ -1,0 +1,7 @@
+package com.BestBot.Core.Service;
+
+public interface CrossoutdbParser {
+    String getApi();
+//    List<Object> testTest();
+
+}
