+++ src/main/resources/application.properties
@@ -15,5 +15,5 @@ spring.datasource.password=112233
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQL82Dialect
 spring.jpa.show-sql=false
 spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.open-in-view=true
